@@ -1,6 +1,7 @@
-module.exports = {
+/** @type {import("prettier").Config} */
+export default {
   singleQuote: true,
-  plugins: [require.resolve('prettier-plugin-astro')],
+  plugins: ['prettier-plugin-astro'],
   overrides: [
     {
       files: '*.astro',
