@@ -1,7 +1,7 @@
+/* eslint-disable @typescript-eslint/triple-slash-reference */
 /// <reference path="../.astro/types.d.ts" />
 /// <reference types="astro/client" />
 
-declare module 'astro-imagetools/components';
 declare module '*.svg?raw' {
   const content: string;
   export default content;
